# api/api-doc.yml
openapi: 3.0.3
info:
  title: Cloudflare 邮箱管理系统 API
  version: 3.0.0
  description: |
    Cloudflare Email Manager (CEM) 的统一 API 规范。系统已精简为 **单管理员模式**，不再开放多用户注册与邮箱申请，所有管理操作均由管理员完成。

    - 所有接口前缀统一为 `/api`
    - 统一使用 HttpOnly Cookie 维持会话
    - localStorage 仅缓存 UI 状态，业务数据一律来自接口
    - `allow_registration` 字段在健康检查与系统配置中继续返回，用于未来扩展（当前永远为只读）
servers:
  - url: http://localhost:8787
    description: 开发环境
  - url: https://cem.xrsec.workers.dev
    description: 生产环境
...
paths:
...
  /api/users/me:
    get:
      tags: [User]
      summary: 获取管理员信息
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 当前管理员信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    put:
      tags: [User]
      summary: 更新管理员信息
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettingsUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /api/forward-rules:
    get:
      tags: [ForwardRules]
      summary: 获取转发规则
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 规则列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForwardRuleListResponse'
        '403': { $ref: '#/components/responses/ForbiddenError' }
    post:
      tags: [ForwardRules]
      summary: 创建转发规则
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForwardRuleInput'
      responses:
        '200': { $ref: '#/components/schemas/ForwardRuleResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/forward-rules/{id}:
    get:
      tags: [ForwardRules]
      summary: 规则详情
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ForwardRuleIdParam'
      responses:
        '200': { $ref: '#/components/schemas/ForwardRuleResponse' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    put:
      tags: [ForwardRules]
      summary: 更新规则
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ForwardRuleIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForwardRuleInput'
      responses:
        '200': { $ref: '#/components/schemas/BasicResponse' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    delete:
      tags: [ForwardRules]
      summary: 删除规则
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ForwardRuleIdParam'
      responses:
        '200': { $ref: '#/components/schemas/BasicResponse' }
        '404': { $ref: '#/components/responses/NotFoundError' }
...
  /api/system/config:
    get:
      tags: [System]
      summary: 获取系统配置
      security:
        - cookieAuth: []
      responses:
        '200': { $ref: '#/components/schemas/SystemConfigResponse' }
    put:
      tags: [System]
      summary: 更新系统配置
      description: 仅管理员可调用；`allow_registration` 为只读，保留未来开关。
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemConfigUpdate'
      responses:
        '200': { $ref: '#/components/schemas/BasicResponse' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_cookies
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    ForwardRuleIdParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
    BasicResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
...
    UserProfileResponse:
      allOf:
        - $ref: '#/components/schemas/BasicResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string, format: email }
        user_type: { type: integer, enum: [1], description: 固定为管理员 }
        created_at: { type: string }
        updated_at: { type: string }
        settings:
          type: object
          properties:
            webhook_url: { type: string, nullable: true }
            webhook_secret: { type: string, nullable: true }
    UserSettingsUpdate:
      type: object
      properties:
        password: { type: string, minLength: 6 }
        webhook_url: { type: string }
        webhook_secret: { type: string }
...
    ForwardRuleInput:
      type: object
      required: [rule_name]
      properties:
        rule_name: { type: string }
        sender_filter: { type: string }
        keyword_filter: { type: string }
        recipient_filter: { type: string }
        enabled: { type: integer, enum: [0, 1] }
        webhooks:
          type: array
          description: 规则关联的多个 webhook 配置
          items:
            $ref: '#/components/schemas/ForwardRuleWebhook'
    ForwardRuleResponse:
      allOf:
        - $ref: '#/components/schemas/BasicResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                rule: { $ref: '#/components/schemas/ForwardRule' }
    ForwardRuleListResponse:
      allOf:
        - $ref: '#/components/schemas/BasicResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                total: { type: integer }
                items:
                  type: array
                  items: { $ref: '#/components/schemas/ForwardRule' }
    ForwardRule:
      type: object
      properties:
        id: { type: integer }
        rule_name: { type: string }
        sender_filter: { type: string }
        keyword_filter: { type: string }
        recipient_filter: { type: string }
        webhook_url: { type: string, description: 保留字段以兼容旧数据 }
        webhook_secret: { type: string, description: 保留字段以兼容旧数据 }
        webhook_type: { type: string, enum: [dingtalk, feishu, bark, custom], description: 保留字段以兼容旧数据 }
        enabled: { type: integer, enum: [0, 1] }
        created_at: { type: string }
        updated_at: { type: string }
        webhooks:
          type: array
          description: 规则关联的多个 webhook 配置
          items:
            $ref: '#/components/schemas/ForwardRuleWebhook'
    ForwardRuleWebhook:
      type: object
      properties:
        id: { type: integer }
        rule_id: { type: integer }
        webhook_url: { type: string }
        webhook_secret: { type: string }
        webhook_type: { type: string, enum: [dingtalk, feishu, bark, custom] }
        enabled: { type: integer, enum: [0, 1] }
        created_at: { type: string }
        updated_at: { type: string }
...
    SystemConfigResponse:
      allOf:
        - $ref: '#/components/schemas/BasicResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                config:
                  type: object
                  additionalProperties: true
                user_role:
                  type: string
                  enum: ['admin']
    SystemConfigUpdate:
      type: object
      properties:
        debug_mode: { type: integer, enum: [0,1] }
        mail_retention_days: { type: integer, description: 邮件保留天数 }
        attachment_retention_days: { type: integer, description: 附件保留天数 }
        max_attachment_size: { type: integer, description: 单位：字节 }
        cookie_max_age: { type: integer, description: 单位：秒 }
        primary_domain: { type: string }
        admin_email: { type: string }
        domains:
          type: array
          items: { type: string }
        default_webhook_url: { type: string, description: 默认 Webhook URL（系统级别，所有邮件都会发送到此 webhook，如果为空则不通知） }
        default_webhook_secret: { type: string, description: 默认 Webhook 密钥（可选） }
        default_webhook_type: { type: string, enum: [dingtalk, feishu, bark, custom], description: 默认 Webhook 类型 }
...
  responses:
    UnauthorizedError:
      description: 未登录或会话失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BasicResponse'
    ForbiddenError:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BasicResponse'
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BasicResponse'
    NotFoundError:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BasicResponse'